Added testing for with and without openssl
